Flow: ReactFiberHotReloading recursive type
